Add Header and AudioPlayer components; update Layout and README for better structure and documentation
