+++ src/App.tsx
@@ -4,7 +4,7 @@ import { TooltipProvider } from '@/components/ui/tooltip';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { AnimatePresence } from 'framer-motion';
-import Layout from './components/Layout';
+import Layout from './components/layout/Layout';
 import Index from './pages/Index';
 import News from './pages/News';
 import About from './pages/About';
